DASH-95 Move server order utils functions to own file
